PaymentPage fixes
